+++ test/unit-tests/expression/node/ArrayNode.test.js
@@ -144,7 +144,7 @@ describe('ArrayNode', function () {
     const c = new ArrayNode([a, b])
 
     assert.throws(function () {
-      c.map(function () {})
+      c.map(function () { return undefined })
     }, /Callback function must return a Node/)
   })