+++ test/unit-tests/expression/node/RelationalNode.test.js
@@ -103,7 +103,7 @@ describe('RelationalNode', function () {
     const n = new RelationalNode(['smaller', 'smaller'], [one, two, three])
 
     assert.throws(function () {
-      n.map(function () {})
+      n.map(function () { return undefined })
     }, /Callback function must return a Node/)
   })