+++ test/unit-tests/function/algebra/rationalize.test.js
@@ -34,6 +34,8 @@ function objToStrings (obj) {
 
 /// ////////////////// rationalize ///////////////////////
 describe('rationalize', function () {
+  this.timeout(10000) // For IE/Edge
+
   it('invalid expression', function () {
     assert.throws(function () { math.rationalize('(x*/2)') }, /Value expected \(char 4\)/)
   })
@@ -115,8 +117,6 @@ describe('rationalize', function () {
   })
 
   it('processing tougher expressions', function () {
-    this.timeout(5000) // For IE/Edge
-
     assert.strictEqual(stri(math.rationalize('2x/(x+2) - x/(x+1)')), 'x^2/(x^2+3*x+2)')
     assert.strictEqual(stri(math.rationalize('2x/( (2x-1) / (3x+2) ) - 5x/ ( (3x+4) / (2x^2-5) ) + 3')),
       '(-20*x^4+28*x^3+104*x^2+6*x-12)/(6*x^2+5*x-4)')
@@ -137,8 +137,6 @@ describe('rationalize', function () {
   })
 
   it('testing complete form', function () {
-    this.timeout(5000) // For IE/Edge
-
     assert.deepStrictEqual(objToStrings(math.rationalize('x+x+x+y', {}, true)), {
       coefficients: '',
       denominator: null,