+++ test/unit-tests/function/matrix/eigs.test.js
@@ -13,11 +13,13 @@ describe('eigs', function () {
     assert.throws(function () { eigs('random') }, /TypeError: Unexpected type of argument/)
     assert.throws(function () { eigs(math.matrix([[1, 2], [2.1, 3]])) }, /Input matrix is not symmetric/)
   })
+
   it('should only accept a matrix with valid element type', function () {
     assert.throws(function () { eigs([['x', 2], [4, 5]]) }, /Mixed matrix element type is not supported/)
     assert.throws(function () { eigs([[1, 2], [2, '5']]) }, /Mixed matrix element type is not supported/)
     assert.throws(function () { eigs([['1', '2'], ['2', '5']]) }, /Matrix element type not supported/)
   })
+
   it('eigenvalue check for diagonal matrix', function () {
     // trivial test
     approx.deepEqual(eigs(
@@ -27,6 +29,7 @@ describe('eigs', function () {
       [[2, 0, 0], [0, 1, 0], [0, 0, 5]]).values, [1, 2, 5]
     )
   })
+
   it('eigenvalue check for 2x2 simple matrix', function () {
     // 2x2 test
     approx.deepEqual(eigs(
@@ -39,6 +42,7 @@ describe('eigs', function () {
       [[5, 2.3], [2.3, 1]]).values, [-0.04795013082563382, 6.047950130825635]
     )
   })
+
   it('eigenvalue check for 3x3 and 4x4 matrix', function () {
     // 3x3 test and 4x4
     approx.deepEqual(eigs(
@@ -55,8 +59,9 @@ describe('eigs', function () {
     [-8.687249803623432, -0.9135495807127523, 2.26552473288741, 5.6502090685149735]
     )
   })
+
   it('eigenvector check', function () {
-    var H = [[-4.78, -1.0, -2.59, -3.26, 4.24, 4.14],
+    const H = [[-4.78, -1.0, -2.59, -3.26, 4.24, 4.14],
       [-1.0, -2.45, -0.92, -2.33, -4.68, 4.27],
       [-2.59, -0.92, -2.45, 4.17, -3.33, 3.05],
       [-3.26, -2.33, 4.17, 2.51, 1.67, 2.24],
@@ -65,13 +70,14 @@ describe('eigs', function () {
     const ans = eigs(H)
     const E = ans.values
     const V = ans.vectors
-    var VtHV = math.multiply(math.transpose(V), H, V)
-    var Ei = Array(H.length)
+    const VtHV = math.multiply(math.transpose(V), H, V)
+    const Ei = Array(H.length)
     for (let i = 0; i < H.length; i++) {
       Ei[i] = VtHV[i][i]
     }
     approx.deepEqual(Ei, E)
   })
+
   it('fractions are supported', function () {
     const aij = math.fraction('1/2')
     approx.deepEqual(eigs(
@@ -81,25 +87,25 @@ describe('eigs', function () {
     [0, 0, 1.5]
     )
   })
+
   it('bigNumber diagonalization is supported', function () {
-    var x = [[math.bignumber(1), math.bignumber(0)], [math.bignumber(0), math.bignumber(1)]]
+    const x = [[math.bignumber(1), math.bignumber(0)], [math.bignumber(0), math.bignumber(1)]]
     approx.deepEqual(eigs(x).values, [math.bignumber(1), math.bignumber(1)])
-    var y = [[math.bignumber(1), math.bignumber(1.0)], [math.bignumber(1.0), math.bignumber(1)]]
-    var E1 = eigs(y).values
+    const y = [[math.bignumber(1), math.bignumber(1.0)], [math.bignumber(1.0), math.bignumber(1)]]
+    const E1 = eigs(y).values
     approx.equal(E1[0].toNumber(), 0.0)
     approx.equal(E1[1].toNumber(), 2.0)
-    var H = [[-4.78, -1.0, -2.59, -3.26, 4.24, 4.14],
+    const H = math.bignumber([[-4.78, -1.0, -2.59, -3.26, 4.24, 4.14],
       [-1.0, -2.45, -0.92, -2.33, -4.68, 4.27],
       [-2.59, -0.92, -2.45, 4.17, -3.33, 3.05],
       [-3.26, -2.33, 4.17, 2.51, 1.67, 2.24],
       [4.24, -4.68, -3.33, 1.67, 2.80, 2.73],
-      [4.14, 4.27, 3.05, 2.24, 2.73, -4.47]]
-    H = math.bignumber(H)
+      [4.14, 4.27, 3.05, 2.24, 2.73, -4.47]])
     const ans = eigs(H)
     const E = ans.values
     const V = ans.vectors
     const VtHV = math.multiply(math.transpose(V), H, V)
-    var Ei = Array(H.length)
+    const Ei = Array(H.length)
     for (let i = 0; i < H.length; i++) {
       Ei[i] = math.bignumber(VtHV[i][i])
     }