+++ test/unit-tests/function/trigonometry/acoth.test.js
@@ -1,3 +1,5 @@
+/* eslint-disable no-loss-of-precision */
+
 import assert from 'assert'
 import math from '../../../../src/defaultInstance.js'
 import approx from '../../../../tools/approx.js'