+++ test/unit-tests/type/matrix/SparseMatrix.test.js
@@ -1379,7 +1379,7 @@ describe('SparseMatrix', function () {
         ptr: [0, 2, 3, 4],
         size: [3, 3]
       })
-      assert.throws(function () { m.map(function () {}, m, true) }, /Cannot invoke map on a Pattern only matrix/)
+      assert.throws(function () { m.map(function () { return undefined }, m, true) }, /Cannot invoke map on a Pattern only matrix/)
     })
   })