+++ HISTORY.md
@@ -29,6 +29,13 @@
     random pick of arrays instead of the elements inside of the nested arrays.
 
 
+# 2020-11-02, version 7.6.0
+
+- Implemented function `rotate(w, theta)`. See #1992, #1160. Thanks @rnd-debug. 
+- Implemented support for custom characters in Units via `Unit.isValidAlpha`. 
+  See #1663, #2000. Thanks @rnd-debug.
+
+
 # 2020-10-10, version 7.5.1
 
 - Fix object pollution vulnerability in `math.config`. Thanks Snyk.